Refactoring: moved events into a folder
